Conflicts:
	dataflowAPI/h/slicing.h
	dataflowAPI/src/slicing.C
	dyninstAPI/src/BPatch.C
	dyninstAPI/src/baseTramp.C
	dyninstAPI/src/emit-x86.C
	dyninstAPI/src/inst-x86.C
	dyninstAPI/src/multiTramp.C
	dyninstAPI/src/registerSpace.C
	dyninstAPI/src/reloc-func.C
	parseAPI/src/IA_IAPI.C
	parseAPI/src/IA_IAPI.h
	parseAPI/src/IA_x86.C
	parseAPI/src/Parser.C
	parseAPI/src/SymtabCodeSource.C

+++ dyninstAPI/src/hybridInstrumentation.C
@@ -70,7 +70,7 @@ static void signalHandlerExitCB_wrapper(BPatch_point *point, void *returnAddr)
 static void synchShadowOrigCB_wrapper(BPatch_point *point, void *toOrig) 
 {
     mal_printf("in synch callback for point 0x%lx toOrig=%d\n",
-               (Address)point->getAddress(), (int)toOrig);
+               (Address)point->getAddress(), (int) (long) toOrig);
 
     BPatch_process *proc = dynamic_cast<BPatch_process*>(
        point->getFunction()->getAddSpace());
